move gamepad logic to a new file
