+++ config.ini
@@ -38,7 +38,6 @@ fmvFps = 20 ; frames per second (seems to match original)
 [gamepad]
 ; axis/button/hat identifiers are expected as SDL id + 1 
 ;                    (1 is the first usable value, 0 to disable)
-axis_dead_zone = 3000;
 axis_yaw = 4 ; roll and yaw axis (initiates turns during flight)
 axis_yaw_inv = false ; false - no inversion needed, true - axis values need inversion
 axis_pitch = 5 ; pitch (moves camera up/down during flight)
@@ -56,8 +55,9 @@ hat_nav_inv = true ; invert vertical axis
 hat_mov = 0 ; longitudinal and transversal movement
 hat_mov_inv = false ; invert vertical axis
 button_spell = 3 ; joystick button-press code used for displaying the spellbook
-button_fire_R = 5 ; 
-button_fire_L = 6 ;
+button_fire_R = 5 ; fire right-hand spell, also right mouse button equivalent
+button_fire_L = 6 ; fire left-hand spell, also left mouse button equivalent
 ;button_fwd = 6 ; joystick button-press code used for forward movement
 ;button_back = 7 ; joystick button-press code used for backward movement
+axis_dead_zone = 3000 ; increase if _mouse_ is unresponsive
