+++ remc2/engine/engine_support.h
@@ -21,6 +21,7 @@
 #include "MapType.h"
 #include "../portability/port_time.h"
 #include "../portability/port_sdl_vga_mouse.h"
+#include "../portability/port_sdl_joystick.h"
 #include "../portability/port_outputs.h"
 #include "../portability/port_show_perifery.h"
